feat: git providers command format
